requirements.txt
pyyaml
tensorflow
numpy